requirements.txt
celery>=4.3.0
Django>=2.2.0
djangorestframework>=3.10.3
redis==3.3.11
